requirements.txt
# Core dependencies for AI customer support assistant
langchain==0.1.0
langchain-openai==0.0.5
openai==1.12.0
flask==3.0.0
python-dotenv==1.0.0
pydantic==2.6.0 